Slime IDLING and WALKING
Had to bring in Slime character as my other idea of using the same character for enemy didn't work. Mainly because of the limited angles it can shoot, and they don't aligned.

Any ways this Slime animation I got from itch.io and it quite good. Let's how it goes.